Scheduler: extend PlatformSpecific class to reduce duplication (#10356)
